script/requirements.txt
bcrypt==5.0.0
pycryptodome==3.23.0
mysql-connector-python==9.5.0
